+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Prior to v1.0.0 this gem was versioned based on the `MAJOR`.`MINOR` version of R
 
 ## Unreleased
 
-- ...
+- Add `Ezcater/RailsTopLevelSqlExecute` to replace `ActiveRecord::Base.connection.execute` with `execute` in `db/migrate/`.
 
 ## v1.1.1